Extended the `KanataLayers` to support every `OsCode` instead of just 256 values, and allow Mouse buttons in `defsrc` (and potentially in the future more `OsCode` values

+++ src/custom_action.rs
@@ -34,7 +34,7 @@ pub enum Btn {
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
 pub struct Coord {
     pub x: u8,
-    pub y: u8,
+    pub y: u16,
 }
 
 #[derive(Debug, Clone, Copy, PartialEq, Eq, Hash)]
